Rejigs structure and naming so it's more in line with standard
This commit moves some files around and renames others to remain consistent with the repo.
